+++ tests/packages/rpath-mixed-layout/first/first.c
@@ -1,0 +1,6 @@
+// SPDX-FileCopyrightText: 2026 The meson-python developers
+//
+// SPDX-License-Identifier: MIT
+
+extern int second(void);
+int first(void) { return 10 + second(); }