+++ tests/packages/rpath-mixed-layout/second/second.c
@@ -1,0 +1,5 @@
+// SPDX-FileCopyrightText: 2026 The meson-python developers
+//
+// SPDX-License-Identifier: MIT
+
+int second(void) { return 2; }